Move icons components to icons folder
