+++ src/components/icons/CloseCartIcon.tsx
