+++ src/components/icons/RemoveCartItemIcon.tsx
