+++ src/components/icons/index.ts
@@ -2,6 +2,7 @@ import AccountIcon from "./AccountIcon";
 import ArrowIcon from "./ArrowIcon";
 import ArrowIcon2 from "./ArrowIcon2";
 import CartIcon from "./CartIcon";
+import CloseCartIcon from "./CloseCartIcon";
 import CloseIcon from "./CloseIcon";
 import CompareIcon from "./CompareIcon";
 import ErrorSnackBarIcon from "./ErrorSnackBarIcon";
@@ -11,6 +12,7 @@ import HeartIcon from "./HeartIcon";
 import LikeIcon from "./LikeIcon";
 import Logo from "./Logo";
 import OverviewIcon from "./OverviewIcon";
+import RemoveCartItemIcon from "./RemoveCartItemIcon";
 import SearchIcon from "./SearchIcon";
 import ShareIcon from "./ShareIcon";
 import SuccessIcon from "./SuccessIcon";
@@ -18,22 +20,24 @@ import ViewIcon from "./ViewIcon";
 import WarningNotifyIcon from "./WarningNotifyIcon";
 
 export {
-  ArrowIcon2,
   AccountIcon,
-  CartIcon,
-  HeartIcon,
-  SearchIcon,
-  CloseIcon,
-  WarningNotifyIcon,
-  SuccessIcon,
-  ErrorSnackBarIcon,
-  ShareIcon,
-  CompareIcon,
-  LikeIcon,
   ArrowIcon,
+  ArrowIcon2,
+  CartIcon,
+  CloseCartIcon,
+  CloseIcon,
+  CompareIcon,
+  ErrorSnackBarIcon,
+  FilterIcon,
+  FourDotIcon,
+  HeartIcon,
+  LikeIcon,
   Logo,
   OverviewIcon,
-  FourDotIcon,
-  FilterIcon,
+  RemoveCartItemIcon,
+  SearchIcon,
+  ShareIcon,
+  SuccessIcon,
   ViewIcon,
+  WarningNotifyIcon,
 };
